Clean up
* Renaming
* Specify Locale of toLowerCase()
* Lowercase Attribute, Property and Tag names after parsing from config
* Extract TestInput Class
