+++ src/test/resources/configWithAdditionalGlobalAndDynamicConditions.xml
